requirements.txt
#pytesseract==0.3.10
easyocr==1.7.1
Pillow==10.3.0
pdf2image==1.16.3
numpy==1.26.4
#torch==2.2.2